renamed document generator main files as index.js
